refactor(trading): moved sell utils
