+++ packages/suite/src/hooks/wallet/trading/form/useTradingSellForm.ts
@@ -6,15 +6,16 @@ import useDebounce from 'react-use/lib/useDebounce';
 
 import { notificationsActions } from '@suite-common/toast-notifications';
 import {
-    TradingSellStepType,
+    type TradingSellStepType,
     type TradingSellType,
-    TradingTransactionSell,
+    type TradingTransactionSell,
     addIdsToQuotes,
     filterQuotesAccordingTags,
     getTradingPaymentMethods,
     getTradingQuotesByPaymentMethod,
     getUnusedAddressFromAccount,
     invityAPI,
+    sellUtils,
     tradingGetSuccessQuotes,
     useTradingInfo,
 } from '@suite-common/trading';
@@ -56,7 +57,7 @@ import {
 } from 'src/types/trading/trading';
 import { TradingSellFormContextProps, TradingSellFormProps } from 'src/types/trading/tradingForm';
 import type { AmountLimitProps } from 'src/utils/suite/validation';
-import { createQuoteLink, getAmountLimits } from 'src/utils/wallet/trading/sellUtils';
+import { createQuoteLink } from 'src/utils/wallet/trading/sellUtils';
 import {
     getTradingCryptoInfo,
     getTradingNetworkDecimals,
@@ -269,7 +270,7 @@ export const useTradingSellForm = ({
                 const currency =
                     cryptoIdToCoinSymbol(quoteRequest.cryptoCurrency) ??
                     quoteRequest.cryptoCurrency;
-                const limits = getAmountLimits({
+                const limits = sellUtils.getAmountLimits({
                     request: quoteRequest,
                     quotes: allQuotes,
                     currency,
