+++ packages/suite/src/views/wallet/trading/common/TradingTransactions/TradingTransaction/TradingTransactionStatus.tsx
@@ -1,12 +1,11 @@
 import { BuyTradeStatus, ExchangeTradeStatus, SellTradeStatus } from 'invity-api';
 import { DefaultTheme, useTheme } from 'styled-components';
 
-import { type TradingTransaction, exchangeUtils } from '@suite-common/trading';
+import { type TradingTransaction, exchangeUtils, sellUtils } from '@suite-common/trading';
 import { Icon, Row, Text } from '@trezor/components';
 
 import { Translation } from 'src/components/suite';
 import { getStatusMessage as getBuyStatusMessage } from 'src/utils/wallet/trading/buyUtils';
-import { getStatusMessage as getSellStatusMessage } from 'src/utils/wallet/trading/sellUtils';
 
 const getBuyTradeData = (status: BuyTradeStatus, theme: DefaultTheme) => {
     const message = getBuyStatusMessage(status);
@@ -42,7 +41,7 @@ const getBuyTradeData = (status: BuyTradeStatus, theme: DefaultTheme) => {
 };
 
 const getSellTradeData = (status: SellTradeStatus, theme: DefaultTheme) => {
-    const message = getSellStatusMessage(status);
+    const message = sellUtils.getStatusMessage(status);
 
     switch (message) {
         case 'TR_SELL_STATUS_PENDING':
